+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [1.4.0]
 
-* Added parameter autoreconnect to automatically reconnect and resubscribe to the websocket topic when an unexpected WebSocket closure occurs.
 * Removed `int WebSocketTimeout` and `int restTimeout` parameters and replaced them with a `Duration timeout` parameter.
+* Added possibility to set a `receiveWindow` parameter that prevents from replay attacks
 * Minor doc changes
 
 ## [1.3.0]